requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
colorlog==6.8.0
typing-extensions>=4.8.0